app now remembers previous user
Initially the app was going to save data about the previoud photes that
were gotten in some config file and then relaod it all at startup. There
is a lot of work involved with that, specifically the work to figure out
from flickr wether there are new photos, so that just getting them all
every time seems the most likely way to do that without bugs.